+++ src/lib/components/token/index.ts
@@ -1,3 +1,4 @@
 export * from "./TokenCard";
 export * from "./TokenComposition";
 export * from "./TokenImageRender";
+export * from "./UnsupportedToken";
